requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pytest
httpx
python-dotenv
azure-identity
azure-keyvault-secrets
